requirements.txt
python-dotenv>=1.0.1
semantic-kernel>=1.22.0
azure-search-documents
fastapi
requests
matplotlib
chainlit
numpy
pandas
openai
pymupdf
pillow
pydantic
json-repair
tiktoken
tenacity
pytest
azure-identity
azure-storage-blob
azure-ai-inference
azure-ai-evaluation
azure-ai-projects 
opentelemetry-instrumentation-openai-v2
opentelemetry-instrumentation-langchain